--- apps/web/ce/store/cycle/index.ts
@@ -1,7 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-export type { ICycleStore } from "@/store/cycle.store";
