+++ apps/web/core/hooks/store/use-analytics.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // types
-import type { IAnalyticsStore } from "@/plane-web/store/analytics.store";
+import type { IBaseAnalyticsStore as IAnalyticsStore } from "@/store/analytics.store";
 
 export const useAnalytics = (): IAnalyticsStore => {
   const context = useContext(StoreContext);
