+++ apps/web/core/hooks/store/use-global-view.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // types
-import type { IGlobalViewStore } from "@/plane-web/store/global-view.store";
+import type { IGlobalViewStore } from "@/store/global-view.store";
 
 export const useGlobalView = (): IGlobalViewStore => {
   const context = useContext(StoreContext);
