+++ apps/web/core/hooks/store/use-power-k.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // types
-import type { IPowerKStore } from "@/plane-web/store/power-k.store";
+import type { IBasePowerKStore as IPowerKStore } from "@/store/base-power-k.store";
 
 export const usePowerK = (): IPowerKStore => {
   const context = useContext(StoreContext);
