+++ apps/web/core/hooks/store/use-project-inbox.ts
@@ -7,7 +7,7 @@
 import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
-import type { IProjectInboxStore } from "@/plane-web/store/project-inbox.store";
+import type { IProjectInboxStore } from "@/store/inbox/project-inbox.store";
 
 export const useProjectInbox = (): IProjectInboxStore => {
   const context = useContext(StoreContext);
