+++ apps/web/core/hooks/store/use-project-state.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // Plane-web
-import type { IStateStore } from "@/plane-web/store/state.store";
+import type { IStateStore } from "@/store/state.store";
 
 export const useProjectState = (): IStateStore => {
   const context = useContext(StoreContext);
